move files to shared directory
