+++ shared/actions/index.js
@@ -2,24 +2,18 @@ export const SET_COUNTER = 'SET_COUNTER'
 export const INCREMENT_COUNTER = 'INCREMENT_COUNTER'
 export const DECREMENT_COUNTER = 'DECREMENT_COUNTER'
 
-export function set(value) {
-  return {
-    type: SET_COUNTER,
-    payload: value
-  }
-}
+export const set = value => ({
+  type: SET_COUNTER,
+  payload: value
+})
 
-export function increment() {
-  return {
-    type: INCREMENT_COUNTER
-  }
-}
+export const increment = () => ({
+  type: INCREMENT_COUNTER
+})
 
-export function decrement() {
-  return {
-    type: DECREMENT_COUNTER
-  }
-}
+export const decrement = () => ({
+  type: DECREMENT_COUNTER
+})
 
 export const incrementIfOdd = () => (dispatch, getState) => {
   const { counter } = getState()
