+++ shared/index.js
@@ -1,0 +1,2 @@
+export { default as routes } from './routes'
+export { default as configureStore } from './store/configureStore'
