+++ shared/reducers/counter.js
@@ -1,7 +1,7 @@
 import { SET_COUNTER, INCREMENT_COUNTER, DECREMENT_COUNTER } from '../actions'
 
-export default function counter(state = 0, action) {
-  
+export const counter = (state = 0, action) => {
+
   switch(action.type) {
     case SET_COUNTER:
       return action.payload
@@ -13,3 +13,5 @@ export default function counter(state = 0, action) {
       return state
   }
 }
+
+export default counter
