+++ shared/reducers/counter.spec.js
