+++ shared/store/configureStore.js
@@ -2,7 +2,7 @@ import { createStore, applyMiddleware } from 'redux'
 import thunk from 'redux-thunk'
 import rootReducer from '../reducers'
 
-export default function configureStore(initialState) {
+const configureStore = initialState => {
   const store = createStore(
     rootReducer,
     initialState,
@@ -19,3 +19,5 @@ export default function configureStore(initialState) {
 
   return store
 }
+
+export default configureStore
