+++ src/App.jsx
@@ -1,8 +1,7 @@
 import React from 'react'
 import { render } from 'react-dom'
 import { Provider } from 'react-redux'
-import configureStore from './store/configureStore'
-import routes from './components/routes'
+import { routes, configureStore } from '../shared'
 
 const initialState = window.__INITIAL_STATE__,
       store = configureStore(initialState),
