+++ src/containers/Home.js
@@ -1,7 +1,7 @@
 import { Home } from '../components/Home'
 import { bindActionCreators } from 'redux'
 import { connect } from 'react-redux'
-import * as CounterActions from '../actions'
+import * as CounterActions from '../../shared/actions'
 
 const mapStateToProps = state => {
   return {
