+++ tests.webpack.js
@@ -1,2 +1,2 @@
-const context = require.context('./src/', true, /.spec\.jsx?$/); //make sure you have your directory and regex test set correctly!
-context.keys().forEach(context);
+const context = require.context('./src/', true, /.spec\.jsx?$/)
+context.keys().map(context)
